Dump split datasets

+++ pedestrians-video-2-carla/src/pedestrians_video_2_carla/skeleton.py
@@ -100,7 +100,7 @@ def main(args):
     model = LitLinearMapper()
 
     # training
-    trainer = pl.Trainer(gpus=1, num_nodes=1, precision=32)
+    trainer = pl.Trainer(gpus=1, num_nodes=1, precision=16)
     trainer.fit(model, datamodule=dm)
